chore: remove macro
