+++ crates/core/src/test_utils/failing_store.rs
@@ -5,13 +5,13 @@ use std::fmt::{Display, Formatter};
 use std::sync::Arc;
 use std::sync::atomic::{AtomicBool, Ordering};
 
+use futures::stream::BoxStream;
 use object_store::memory::InMemory;
 use object_store::path::Path;
 use object_store::{
-    MultipartUpload, PutMultipartOptions, PutOptions, PutPayload, PutResult, UploadPart,
+    CopyOptions, GetOptions, GetResult, ListResult, MultipartUpload, ObjectMeta, ObjectStore,
+    PutMultipartOptions, PutOptions, PutPayload, PutResult, UploadPart,
 };
-
-use super::impl_object_store_delegating_reads;
 
 /// Wraps [`InMemory`], but every multipart upload it hands out fails `put_part`
 /// and `complete`, and records whether `abort` was called on it. Plain `put`s
@@ -62,29 +62,60 @@ impl Display for FailingMultipartStore {
     }
 }
 
-impl_object_store_delegating_reads! {
-    for FailingMultipartStore {
-        async fn put_opts(
-            &self,
-            location: &Path,
-            payload: PutPayload,
-            opts: PutOptions,
-        ) -> object_store::Result<PutResult> {
-            self.inner.put_opts(location, payload, opts).await
-        }
+#[async_trait::async_trait]
+impl ObjectStore for FailingMultipartStore {
+    async fn put_opts(
+        &self,
+        location: &Path,
+        payload: PutPayload,
+        opts: PutOptions,
+    ) -> object_store::Result<PutResult> {
+        self.inner.put_opts(location, payload, opts).await
+    }
 
-        async fn put_multipart_opts(
-            &self,
-            _location: &Path,
-            _opts: PutMultipartOptions,
-        ) -> object_store::Result<Box<dyn MultipartUpload>> {
-            if self.fail_multipart_create.load(Ordering::Acquire) {
-                return Err(upload_failure());
-            }
-            self.multipart_started.store(true, Ordering::Release);
-            Ok(Box::new(FailingUpload {
-                aborted: self.multipart_aborted.clone(),
-            }))
+    async fn put_multipart_opts(
+        &self,
+        _location: &Path,
+        _opts: PutMultipartOptions,
+    ) -> object_store::Result<Box<dyn MultipartUpload>> {
+        if self.fail_multipart_create.load(Ordering::Acquire) {
+            return Err(upload_failure());
         }
+        self.multipart_started.store(true, Ordering::Release);
+        Ok(Box::new(FailingUpload {
+            aborted: self.multipart_aborted.clone(),
+        }))
+    }
+
+    async fn get_opts(
+        &self,
+        location: &Path,
+        options: GetOptions,
+    ) -> object_store::Result<GetResult> {
+        self.inner.get_opts(location, options).await
+    }
+
+    fn delete_stream(
+        &self,
+        locations: BoxStream<'static, object_store::Result<Path>>,
+    ) -> BoxStream<'static, object_store::Result<Path>> {
+        self.inner.delete_stream(locations)
+    }
+
+    fn list(&self, prefix: Option<&Path>) -> BoxStream<'static, object_store::Result<ObjectMeta>> {
+        self.inner.list(prefix)
+    }
+
+    async fn list_with_delimiter(&self, prefix: Option<&Path>) -> object_store::Result<ListResult> {
+        self.inner.list_with_delimiter(prefix).await
+    }
+
+    async fn copy_opts(
+        &self,
+        from: &Path,
+        to: &Path,
+        options: CopyOptions,
+    ) -> object_store::Result<()> {
+        self.inner.copy_opts(from, to, options).await
     }
 }
